refactor : 자료형 long으로 변경

+++ src/main/java/com/focusup/domain/routine/service/RoutineServiceImpl.java
@@ -133,8 +133,8 @@ public class RoutineServiceImpl implements RoutineService{
         // 유저의 level successcount 확인
         LevelHistory levelHistory = levelHistoryRepository.findByUser(user);
 
-        int boostCount = (int) execTime / levelHistory.getLevel().getMinute(); // level에 따란 boost time 변경
-        levelHistory.addSuccessCount(boostCount);
+        long boostCount = execTime / levelHistory.getLevel().getMinute(); // level에 따란 boost time 변경
+        levelHistory.addSuccessCount((int)boostCount);
 
         // boostCount가 5 이상이면 레벨 업
         while (levelHistory.getSuccessCount() > 4) {
